// template/react-starter-kit/src/.generated/api/model/supportControllerGetSupportTicketsParams.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */
import type { SortDirection } from './sortDirection';
import type { SupportControllerGetSupportTicketsSortItem } from './supportControllerGetSupportTicketsSortItem';
import type { SupportTicketPriority } from './supportTicketPriority';
import type { SupportTicketStatus } from './supportTicketStatus';

export type SupportControllerGetSupportTicketsParams = {
sort?: SupportControllerGetSupportTicketsSortItem[];
direction?: SortDirection[];
search?: string;
page?: number;
/**
 * @maximum 100
 */
limit?: number;
status?: SupportTicketStatus;
priority?: SupportTicketPriority;
category?: string;
};