// template/react-starter-kit/src/.generated/api/model/supportControllerGetSupportTicketsSortItem.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Nest Starter Kit
 * NestJS + MikroORM Starter Kit API
 * OpenAPI spec version: 1.0.0
 */

export type SupportControllerGetSupportTicketsSortItem = typeof SupportControllerGetSupportTicketsSortItem[keyof typeof SupportControllerGetSupportTicketsSortItem];


export const SupportControllerGetSupportTicketsSortItem = {
  category: 'category',
  title: 'title',
  status: 'status',
  priority: 'priority',
  createdAt: 'createdAt',
  updatedAt: 'updatedAt',
  id: 'id',
} as const;